Avoid backtracking in FusionOptimizer
The change in number of fused kernels has to do with the order of iteration, and could be replicated in the old approach by iterating in topological order. It was an accident that it happen to visit in an order where it found two connected branches, instead of a single one. The underlying limitation already existed and is described in https://github.com/pymc-devs/pytensor/issues/249